requirements.txt
openai
pyttsx3
SpeechRecognition
whisper-openai
